Version bump compiz to 0.9.7.8
The set is bumped to the last versions of each package: ccsm, compizconfig-python and libcompizconfig to 0.9.5.92;
compiz-plugins-main to 0.9.7.2 and compiz-plugins-extra to 0.9.7.0.
Remove compiz-plugins-unsupported until a new tarball is released.
The change in the naming scheme done in 0.9.4 is reverted, meaning compiz-core is compiz and
ccsm, compizconfig-python and libcompizconfig lose the compiz- preffix.
Developement has moved to launchpad.
Nuke the individual exlibs.
Fix dependencies.
Rename set to compiz-set to avoid confusions.